Fix node counting bug for certain balanced trees
If tree was constructed to be fully balanced the count empty nodes function would skip the child nodes of empty non-leaf nodes. This issue doesn't really have anything to do with the balance status of a tree, but if you don't construct a tree with the balanced requirement, you usually stop recursion when encountering an empty node, as all children of that node would also be empty.