Octree rendering working (partially)
My renderer shows similar results to the one I based it off, however it is currently not traversing correctly. It basically always adds the entire root node and all other nodes (internal AND children) that are relevant, and then ray marches through all these, making it slower than usual. To fix this, I am going to implement a stack that traverses deeper into the tree until root nodes are found.