Fix critical bounds error in setMinMaxVal()
When sampling for min/max values in the volume for all nodes, the sampling always started from 0 and went up to the node's max corner. This took a heavy toll on the octree performance and introduced undesired intersection behavior. Now the sampling is done in the correct interval from the node's min corner to the node's max corner. There has also been some minor adjustment to code structure in the shader to make it easier to adjust impactful parameters.